feature: add interfaces to make the services more decoupled
