+++ src/main/java/br/com/biblioteca/services/AuthService.java
@@ -1,0 +1,8 @@
+package br.com.biblioteca.services;
+
+
+import br.com.biblioteca.dtos.LoginDto;
+
+public interface AuthService  {
+    String login(LoginDto loginDto);
+}
